added const Edit and List
