+++ admin/src/post-edit.jsx
@@ -13,21 +13,6 @@ export const PostEdit = () => (
   </Edit>
 );
 
-export const TeacherEdit = () => (
-  <Edit title="Edit teachers">
-    <SimpleForm>
-      <TextInput disabled source="id" />
-      <TextInput source="firstname" />
-      <TextInput source="lastname" />
-      <TextInput source="initials" />
-      <TextInput source="phone" />
-      <TextInput label="Email Address" source="email" type="email" />
-      <BooleanInput label="hide" source="hide" />
-      <PasswordInput source="password" />
-      <TextInput source="roles" />
-    </SimpleForm>
-  </Edit>
-);
 
 export const SchoolEdit = () => (
   <Edit title="Edit schools">
