--- admin/src/users.jsx
@@ -1,16 +1,0 @@
-import { Datagrid, EmailField, List, TextField, UrlField } from 'react-admin';
-
-export const UserList = () => (
-  <List>
-    <Datagrid rowClick="edit">
-      <TextField source="id" />
-      <TextField source="name" />
-      <TextField source="username" />
-      <EmailField source="email" />
-      <TextField source="address.street" />
-      <TextField source="phone" />
-      <UrlField source="website" />
-      <TextField source="company.name" />
-    </Datagrid>
-  </List>
-);
